+++ app/views/calendars/show.html.erb
@@ -5,7 +5,9 @@
 
 <%= content_tag :div, class: "simple-calendar #{session[:calendar_color]}" do %>
   <%= month_calendar events: @events do |date, events| %>
-    <div class="calendar-date"><%= date.day %></div>
+    <div class="calendar-date">
+      <%= link_to date.day, new_schedule_path(date: date, calendar_id: @calendar.id), class: 'date-link' %>
+    </div>
     <div class="event-list">
       <% events.each do |event| %>
         <div class="event-item"><%= event.title %></div>
