+++ app/views/schedules/edit.html.erb
@@ -1,9 +1,9 @@
-<h2>New Schedule for <%= params[:date] %></h2>
+<h2>Edit Schedule</h2>
 
 <%= form_with(model: @schedule, local: true) do |form| %>
   <div class="field">
     <%= form.label :date %>
-    <%= form.date_field :date, value: params[:date] %>
+    <%= form.date_field :date %>
   </div>
 
   <div class="field">
@@ -20,8 +20,6 @@
     <%= form.label :duration, "Duration (seconds)" %>
     <%= form.number_field :duration %>
   </div>
-
-  <%= hidden_field_tag :calendar_id, params[:calendar_id] %>
 
   <div class="actions">
     <%= form.submit %>
