+++ db/migrate/20240522171655_add_details_to_schedules.rb
@@ -1,0 +1,6 @@
+class AddDetailsToSchedules < ActiveRecord::Migration[6.1]
+  def change
+    add_column :schedules, :repetitions, :integer
+    add_column :schedules, :duration, :integer
+  end
+end
