+++ db/schema.rb
@@ -10,7 +10,7 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2024_05_21_124613) do
+ActiveRecord::Schema.define(version: 2024_05_22_171655) do
 
   # These are extensions that must be enabled in order to support this database
   enable_extension "plpgsql"
@@ -89,6 +89,8 @@ ActiveRecord::Schema.define(version: 2024_05_21_124613) do
     t.date "date", null: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
+    t.integer "repetitions"
+    t.integer "duration"
     t.index ["calendar_id"], name: "index_schedules_on_calendar_id"
     t.index ["exercise_id"], name: "index_schedules_on_exercise_id"
   end
